login: respond with CORS headers to forward'd auth calls triggered from pre-flight requests
Our LB setup uses "forward auth" which means traefik issues an internal
HTTP request to this method that checks each request for a valid cookie.
We use this instead of requiring each endpoint to be aware of auth.

However, when a pre-flight request comes through that also triggers an
internal forward auth call into this method, but without the cookie.

Thus, we need to check if the request was forwarded as a pre-flight request
and if so just respond with 200 and our usual CORS headers.